+++ Sources/KeyServer/LoginService/KeyServerApi+WalletDataProvider.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension KeyServerApi: KeyServerLoginService.WalletDataProvider {
+    
+    public func walletData(
+        walletId: String,
+        completion: @escaping (Result<WalletDataResponse, Error>) -> Void
+    ) {
+        
+        getWallet(
+            walletId: walletId,
+            completion: completion
+        )
+    }
+}