+++ Sources/KeyServer/LoginService/KeyServerLoginService.WalletKDFProvider.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+public protocol KeyServerLoginServiceWalletKDFProvider {
+    
+    func walletKDF(
+        login: String,
+        isRecovery: Bool,
+        completion: @escaping (_ result: Swift.Result<WalletKDFParams, Swift.Error>) -> Void
+    )
+}
+
+public extension KeyServerLoginService {
+    
+    typealias WalletKDFProvider = KeyServerLoginServiceWalletKDFProvider
+}